put sql in viewsModel.py and put all models in models folder
